move TLS proxy logic into a module
